# docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Untripical API",
        "description": "API aplication Untripical",
        "contact": {
            "name": "Mateusz",
            "email": "mateusz@untripical.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Untripical server"
        }
    ],
    "paths": {
        "/user/update": {
            "put": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "updateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserUpdateResponseWithTokenDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tripStop/{orderIndex}/{tripPlanId}": {
            "put": {
                "tags": [
                    "trip-stop-controller"
                ],
                "operationId": "updateTripStop",
                "parameters": [
                    {
                        "name": "orderIndex",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "tripPlanId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripStopUpdateRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripStopResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "trip-stop-controller"
                ],
                "operationId": "deleteTripStop",
                "parameters": [
                    {
                        "name": "orderIndex",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "tripPlanId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripStopResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/plan/update/{name}": {
            "put": {
                "tags": [
                    "trip-plan-controller"
                ],
                "operationId": "updateTripPlan",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripPlanRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripPlanResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/register": {
            "post": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tripStop": {
            "post": {
                "tags": [
                    "trip-stop-controller"
                ],
                "operationId": "addTripStop",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripStopRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripStopResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/plan": {
            "post": {
                "tags": [
                    "trip-plan-controller"
                ],
                "operationId": "addTripPlan",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripPlanRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripPlanResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tripStop/{tripPlanId}": {
            "get": {
                "tags": [
                    "trip-stop-controller"
                ],
                "operationId": "getAllTripStopsWithPlanId",
                "parameters": [
                    {
                        "name": "tripPlanId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TripStopResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/plan/{name}": {
            "get": {
                "tags": [
                    "trip-plan-controller"
                ],
                "operationId": "getTripPlanByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripPlanResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "trip-plan-controller"
                ],
                "operationId": "deleteTripPlan",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripPlanResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/plan/date": {
            "get": {
                "tags": [
                    "trip-plan-controller"
                ],
                "operationId": "getTripPlanByDate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripPlanRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TripPlanResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/{id}": {
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "deleteUserOrGuideByAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserUpdateDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserResponseDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "userRole": {
                        "type": "string",
                        "enum": [
                            "USER",
                            "GUIDE",
                            "ADMIN"
                        ]
                    }
                }
            },
            "UserUpdateResponseWithTokenDTO": {
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserResponseDTO"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
            "TripStopUpdateRequestDTO": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    }
                }
            },
            "TripStopResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "distanceToNext": {
                        "type": "number",
                        "format": "double"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "TripPlanRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "dayWhenTripPlanIsStarting": {
                        "type": "string",
                        "format": "date",
                        "minLength": 1
                    }
                }
            },
            "TripPlanResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "assignedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "dayWhenTripPlanIsStarting": {
                        "type": "string",
                        "format": "date"
                    },
                    "tripStops": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TripStopResponseDTO"
                        }
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "tripPlanId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalDistance": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "UserRegisterRequestDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                }
            },
...
            "Place": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "city": {
                        "type": "string",
                        "minLength": 1
                    },
                    "addressStreet": {
                        "type": "string",
                        "minLength": 1
                    },
                    "addressNumber": {
                        "type": "string",
                        "minLength": 1
                    },
                    "placeType": {
                        "type": "string",
                        "enum": [
                            "TYPICAL",
                            "UNTYPICAL"
                        ]
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "photoUrl": {
                        "type": "string",
                        "minLength": 1
                    },
                    "avgRating": {
                        "type": "number",
                        "format": "double"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "WAITING_FOR_APPROVAL",
                            "APPROVED",
                            "REJECTED"
                        ]
                    },
                    "tripStops": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TripStop"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    },
                    "region": {
                        "$ref": "#/components/schemas/Region"
                    },
                    "announcements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Announcement"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "required": [
                    "avgRating",
                    "isActive",
                    "placeType",
                    "postalCode",
                    "status"
                ]
            },
...
            "Review": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "numberOfStars": {
                        "type": "number",
                        "format": "double",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "place": {
                        "$ref": "#/components/schemas/Place"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "guideDetails": {
                        "$ref": "#/components/schemas/GuideDetails"
                    }
                },
                "required": [
                    "createdAt",
                    "isActive",
                    "numberOfStars",
                    "orderIndex"
                ]
            },
            "TripPlan": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "assignedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "dayWhenTripPlanIsStarting": {
                        "type": "string",
                        "format": "date"
                    },
                    "totalDistance": {
                        "type": "number",
                        "format": "double"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "required": [
                    "assignedAt",
                    "dayWhenTripPlanIsStarting",
                    "isActive"
                ]
            },
            "TripStop": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "distanceToNext": {
                        "type": "number",
                        "format": "double"
                    },
                    "tripPlan": {
                        "$ref": "#/components/schemas/TripPlan"
                    }
                },
                "required": [
                    "orderIndex"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "userRole": {
                        "type": "string",
                        "enum": [
                            "USER",
                            "GUIDE",
                            "ADMIN"
                        ]
                    },
                    "places": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Place"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    },
                    "tripPlans": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TripPlan"
                        }
                    }
                },
                "required": [
                    "createdAt",
                    "isActive",
                    "userRole"
                ]
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "TripStopRequestDTO": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "tripPlanId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "placeId",
                    "tripPlanId"
                ]
            },
...
        }
    }
}